use proper function in TrtottleManager for localtime ( localtime_r) some fixes around custom share some fixes around getLocalIp some cleanup

+++ dcpp/ShareManager.cpp
@@ -63,9 +63,9 @@ atomic_flag ShareManager::refreshing = ATOMIC_FLAG_INIT;
 ShareManager::ShareManager(string _name) : hits(0), xmlListLen(0), bzXmlListLen(0),
 	xmlDirty(true), forceXmlRefresh(true), refreshDirs(false), update(false), listN(0),
 	lastXmlUpdate(0), lastFullUpdate(GET_TICK()), bloom(1<<20), bzXmlRoot(NULL),xmlRoot(NULL),
-	name(_name)
+	id(_name)
 {
-	if(name.empty())
+	if(id.empty())
 		SettingsManager::getInstance()->addListener(this);
 	
 	TimerManager::getInstance()->addListener(this);
@@ -93,10 +93,24 @@ ShareManager::~ShareManager() {
 
 	join();
 
+	/*if(bzXmlRef.get()) {
+		bzXmlRef.reset();
+		File::deleteFile(getBZXmlFile());
+	}*/
+}
+
+//Borowed from wxStrong for Custom Share
+void ShareManager::setName(const string& _id) {
+	if(_id == id)
+		return; // nothing to change
+
 	if(bzXmlRef.get()) {
+		// delete list with old name
 		bzXmlRef.reset();
 		File::deleteFile(getBZXmlFile());
 	}
+
+	id = _id;
 }
 
 ShareManager::Directory::Directory(const string& aName, const ShareManager::Directory::Ptr& aParent) :
@@ -122,11 +136,11 @@ string ShareManager::Directory::getFullName() const noexcept {
 	return getParent()->getFullName() + getName() + '\\';
 }
 
-string ShareManager::Directory::getRealPath(const std::string& path) const {
+string ShareManager::Directory::getRealPath(const ShareManager* manager,const std::string& path) const {
 	if(getParent()) {
-		return getParent()->getRealPath(getRealName() + PATH_SEPARATOR_STR + path);
+		return getParent()->getRealPath(manager,getRealName() + PATH_SEPARATOR_STR + path);
 	} else {
-		return ShareManager::getInstance()->findRealRoot(getRealName(), path);
+		return /*ShareManager::getInstance()*/manager->findRealRoot(getRealName(), path);
 	}
 }
 
@@ -192,7 +206,7 @@ pair<string, int64_t> ShareManager::toRealWithSize(const string& virtualFile, bo
 	}
 
 	auto f = findFile(virtualFile);
-	return make_pair(f.getRealPath(), f.getSize());
+	return make_pair(f.getRealPath(this), f.getSize());
 }
 
 StringList ShareManager::getRealPaths(const string& virtualPath) {
@@ -209,7 +223,7 @@ StringList ShareManager::getRealPaths(const string& virtualPath) {
 
 		// imitate Directory::getRealPath
 		if(d->getParent()) {
-			ret.push_back(d->getParent()->getRealPath(d->getName()));
+			ret.push_back(d->getParent()->getRealPath(this,d->getName()));
 		} else {
 			for(auto& i: shares) {
 				if(Util::stricmp(i.second, d->getName()) == 0) {
@@ -452,7 +466,7 @@ void ShareManager::merge(const Directory::Ptr& directory, const string& realPath
 	auto i = directories.find(directory->getName());
 	if(i != directories.end()) {
 		dcdebug("Merging directory <%s> into %s\n", realPath.c_str(), directory->getName().c_str());
-		i->second->merge(directory, realPath);
+		i->second->merge(this,directory, realPath);
 
 	} else {
 		dcdebug("Adding new directory %s\n", directory->getName().c_str());
@@ -460,7 +474,7 @@ void ShareManager::merge(const Directory::Ptr& directory, const string& realPath
 	}
 }
 
-void ShareManager::Directory::merge(const Directory::Ptr& source, const string& realPath) {
+void ShareManager::Directory::merge(const ShareManager* manager,const Directory::Ptr& source, const string& realPath) {
 	// merge directories
 	for(auto& i: source->directories) {
 		auto subSource = i.second;
@@ -474,13 +488,13 @@ void ShareManager::Directory::merge(const Directory::Ptr& source, const string& 
 			auto f = findFile(subSource->getName());
 			if(f != files.end()) {
 				// we have a file that has the same name as the dir being merged; rename it.
-				const_cast<File&>(*f).validateName(Util::getFilePath(f->getRealPath()));
+				const_cast<File&>(*f).validateName(Util::getFilePath(f->getRealPath(manager)));
 			}
 
 		} else {
 			// the directory was already existing; merge into it.
 			auto subTarget = ti->second;
-			subTarget->merge(subSource, realPath + subSource->getName() + PATH_SEPARATOR);
+			subTarget->merge(manager,subSource, realPath + subSource->getName() + PATH_SEPARATOR);
 		}
 	}
 
@@ -812,7 +826,7 @@ void ShareManager::updateIndices(Directory& dir, const decltype(std::declval<Dir
 	} else {
 		if(!SETTING(LIST_DUPES)) {
 			try {
-				LogManager::getInstance()->message(_("Duplicate file will not be shared: ")+Util::addBrackets(f.getRealPath())+_("Size: ")+Util::toString(f.getSize())+_(" B) Dupe matched against: ") + Util::addBrackets(j->second->getRealPath()));
+				LogManager::getInstance()->message(_("Duplicate file will not be shared: ")+Util::addBrackets(f.getRealPath(this))+_("Size: ")+Util::toString(f.getSize())+_(" B) Dupe matched against: ") + Util::addBrackets(j->second->getRealPath(this)));
 				dir.files.erase(i);
 			} catch (const ShareException&) {
 			}
